+++ assets/js/components/sidebar.vue
@@ -5,24 +5,20 @@
         <h5 class="text-center">
             Categories
         </h5>
+
         <ul class="nav flex-column mb4">
-            <li class="nav-item">
+            <!-- key="index" is an unique key for Vue internal tracking -->
+            <li
+                v-for="(category, index) in categories"
+                :key="index"
+                class="nav-item"
+            >
                 <a
+                    :href="category.link"
                     class="nav-link"
-                    href="/"
-                >All Products</a>
-            </li>
-            <li class="nav-item">
-                <a
-                    class="nav-link"
-                    href="#"
-                >Category A</a>
-            </li>
-            <li class="nav-item">
-                <a
-                    class="nav-link"
-                    href="#"
-                >Category B</a>
+                >
+                    {{ category.name }} <!-- dynamic content -->
+                </a>
             </li>
         </ul>
     </div>
@@ -31,6 +27,14 @@
 <script>
 export default {
     name: 'Sidebar',
+    data() {
+        return {
+            categories: [
+                { name: 'Dot Matrix Printers', link: '#' },
+                { name: 'Iomega Zip Drives', link: '#' },
+            ],
+        };
+    },
 };
 </script>