+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # The Delightful World of Vue
 ## [07.02.2025]
+- Communication UP with $emit and using Shared CSS
+- Where should a Piece of Data Live?
 - Computed Properties
 - v-if, v-show and Conditional Classes
 - Reactivity